+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 _Note: Numbers like (\#123) point to closed Pull Requests on the fractal-web repository._
 
-# Unreleased
+# 1.27.7
 
 - Support starring datasets (\#1100).
+- Use new combined latest-job & status endpoint (\#1103).
+- Drop obsolete task-group properties (`venv_size_in_kB` and `venv_file_number`) (\#1103).
 
 # 1.27.6